ALUMNOS/MDAA/MAR_SOLER/AWS_ALMACENAMIENTO/MAR/requirements.txt
dotenv
psycopg2-binary
faker
pandas